+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.23.6] 2026/02/04
+
+### Changed
+- Extracted dashboard aggregation into `Dashboard::IndexData`.
+- Extracted reports aggregation into `Reports::IndexData`.
+
 ## [1.23.5] 2026/02/04
 
 ### Changed